Move exceptions to separate classes.
