+++ src/main/java/com/leverx/learningmanagementsystem/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package com.leverx.learningmanagementsystem.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
