+++ src/main/java/com/leverx/learningmanagementsystem/exception/IncorrectResultSizeException.java
@@ -1,0 +1,7 @@
+package com.leverx.learningmanagementsystem.exception;
+
+public class IncorrectResultSizeException extends RuntimeException {
+    public IncorrectResultSizeException(String message) {
+        super(message);
+    }
+}
