+++ src/main/java/com/leverx/learningmanagementsystem/exception/InvalidCourseDatesException.java
@@ -1,0 +1,7 @@
+package com.leverx.learningmanagementsystem.exception;
+
+public class InvalidCourseDatesException extends RuntimeException {
+    public InvalidCourseDatesException(String message) {
+        super(message);
+    }
+}
