+++ src/main/java/com/leverx/learningmanagementsystem/exception/NotEnoughCoinsException.java
@@ -1,0 +1,7 @@
+package com.leverx.learningmanagementsystem.exception;
+
+public class NotEnoughCoinsException extends RuntimeException {
+    public NotEnoughCoinsException(String message) {
+        super(message);
+    }
+}
