+++ src/main/java/com/leverx/learningmanagementsystem/exception/StudentAlreadyEnrolledException.java
@@ -1,0 +1,7 @@
+package com.leverx.learningmanagementsystem.exception;
+
+public class StudentAlreadyEnrolledException extends RuntimeException {
+    public StudentAlreadyEnrolledException(String message) {
+        super(message);
+    }
+}
