+++ src/main/java/com/leverx/learningmanagementsystem/exception/handler/GlobalExceptionHandler.java
@@ -1,14 +1,13 @@
 package com.leverx.learningmanagementsystem.exception.handler;
 
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.IncorrectResultSizeException;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.InvalidCourseDatesException;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.EntityNotFoundException;
+import com.leverx.learningmanagementsystem.exception.EntityNotFoundException;
+import com.leverx.learningmanagementsystem.exception.IncorrectResultSizeException;
+import com.leverx.learningmanagementsystem.exception.InvalidCourseDatesException;
 import com.leverx.learningmanagementsystem.exception.response.ErrorResponse;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -16,32 +15,36 @@ import org.springframework.web.method.annotation.MethodArgumentTypeMismatchExcep
 
 import static com.leverx.learningmanagementsystem.utils.DataFormatUtils.DATE_FORMAT;
 import static com.leverx.learningmanagementsystem.utils.DataFormatUtils.DATE_TIME_FORMAT;
-import static org.springframework.http.HttpStatus.*;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+import static org.springframework.http.HttpStatus.CONFLICT;
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
+
+
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
     @ExceptionHandler(EntityNotFoundException.class)
     @ResponseStatus(NOT_FOUND)
-    public @ResponseBody ErrorResponse handleEntityNotFoundException(EntityNotFoundException ex) {
+    public ErrorResponse handleEntityNotFoundException(EntityNotFoundException ex) {
         return new ErrorResponse(NOT_FOUND.value(), ex.getMessage());
     }
 
     @ExceptionHandler(IncorrectResultSizeException.class)
     @ResponseStatus(CONFLICT)
-    public @ResponseBody ErrorResponse handleMismatchException(IncorrectResultSizeException ex) {
+    public ErrorResponse handleMismatchException(IncorrectResultSizeException ex) {
         return new ErrorResponse(CONFLICT.value(), ex.getMessage());
     }
 
     @ExceptionHandler(InvalidCourseDatesException.class)
     @ResponseStatus(CONFLICT)
-    public @ResponseBody ErrorResponse handleInvalidCourseDatesException(InvalidCourseDatesException ex) {
+    public ErrorResponse handleInvalidCourseDatesException(InvalidCourseDatesException ex) {
         return new ErrorResponse(CONFLICT.value(), ex.getMessage());
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
     @ResponseStatus(BAD_REQUEST)
-    public @ResponseBody ErrorResponse handleValidationException(MethodArgumentNotValidException ex) {
+    public ErrorResponse handleValidationException(MethodArgumentNotValidException ex) {
         String errorMessage = ex.getBindingResult()
                 .getFieldErrors()
                 .stream()
@@ -53,7 +56,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(HttpMessageNotReadableException.class)
     @ResponseStatus(BAD_REQUEST)
-    public @ResponseBody ErrorResponse handleMessageNotReadableException(HttpMessageNotReadableException ex) {
+    public ErrorResponse handleMessageNotReadableException(HttpMessageNotReadableException ex) {
         Throwable cause = ex.getCause();
         if (cause instanceof InvalidFormatException &&
                 ((InvalidFormatException) cause).getTargetType() == java.util.UUID.class) {
@@ -73,7 +76,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(MethodArgumentTypeMismatchException.class)
     @ResponseStatus(BAD_REQUEST)
-    public @ResponseBody ErrorResponse handleMethodArgumentTypeMismatch(MethodArgumentTypeMismatchException ex) {
+    public ErrorResponse handleMethodArgumentTypeMismatch(MethodArgumentTypeMismatchException ex) {
         if (ex.getRequiredType() == java.util.UUID.class) {
             return new ErrorResponse(BAD_REQUEST.value(), "Invalid UUID format: " + ex.getValue());
         }
@@ -83,13 +86,13 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(RuntimeException.class)
     @ResponseStatus(NOT_FOUND)
-    public @ResponseBody ErrorResponse handleRuntimeException(RuntimeException ex) {
+    public ErrorResponse handleRuntimeException(RuntimeException ex) {
         return new ErrorResponse(NOT_FOUND.value(), ex.getMessage());
     }
 
     @ExceptionHandler(Exception.class)
     @ResponseStatus(NOT_FOUND)
-    public @ResponseBody ErrorResponse handleException(Exception ex) {
+    public ErrorResponse handleException(Exception ex) {
         return new ErrorResponse(NOT_FOUND.value(), ex.getMessage());
     }
 }
