+++ src/main/java/com/leverx/learningmanagementsystem/service/impl/StudentServiceImpl.java
@@ -3,10 +3,10 @@ package com.leverx.learningmanagementsystem.service.impl;
 import com.leverx.learningmanagementsystem.dto.student.CreateStudentDto;
 import com.leverx.learningmanagementsystem.entity.Course;
 import com.leverx.learningmanagementsystem.entity.Student;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.StudentAlreadyEnrolledException;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.EntityNotFoundException;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.IncorrectResultSizeException;
-import com.leverx.learningmanagementsystem.exception.EntityValidationException.NotEnoughCoinsException;
+import com.leverx.learningmanagementsystem.exception.EntityNotFoundException;
+import com.leverx.learningmanagementsystem.exception.IncorrectResultSizeException;
+import com.leverx.learningmanagementsystem.exception.NotEnoughCoinsException;
+import com.leverx.learningmanagementsystem.exception.StudentAlreadyEnrolledException;
 import com.leverx.learningmanagementsystem.mapper.student.StudentMapper;
 import com.leverx.learningmanagementsystem.repository.CourseRepository;
 import com.leverx.learningmanagementsystem.repository.StudentRepository;
@@ -17,9 +17,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 
 @Slf4j
